refactor(cli): extract source-catalog cache cluster + shared server-types
#354 pass 2. Moves the source-session catalog cache cluster out of server.ts
into server-source-catalog.ts (build/normalize/merge/update + freshness probes
+ getStaleSourceProviders), and introduces server-types.ts to hold the shared
record/cache types (SourceSummaryRecord, ReplaySummary, CachedSourceRecord,
ProviderDiscoveryState, SourceSessionCatalogCache, NormalizedSource...Cache,
SourceProviderFreshnessProbe).

- server-source-catalog.ts and server-enrichment.ts depend on server-types.ts,
  not on server.ts — no reverse dependency.
- server.ts re-exports SourceSummaryRecord so existing test imports
  (../src/server.js) keep working; __testables re-imports the moved functions.
- Dropped now-unused imports (getPiSessionsDir, FileCacheEntry) from server.ts.

server.ts: 3573 -> 3263 lines. No behavior change — all 340 CLI tests pass,
typecheck + lint clean.

Refs #354
